requirements.txt
django>=4.2
azure-storage-blob>=12.19.0
azure-data-tables>=12.6.0
gunicorn==23.0.0
